requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.8
requests==2.32.3
RPi.GPIO==0.7.1
smbus2==0.4.3
urllib3==2.2.2
